requirements.txt
beautifulsoup4==4.14.3
faiss==1.5.3
langchain_community==0.4.1
langchain_openai==1.1.10
langchain_text_splitters==1.1.1
numpy==2.4.2
pandas==3.0.1
python-dotenv==1.2.1
Requests==2.32.5
sentence_transformers==5.2.3
streamlit==1.54.0
